requirements.txt
pandas==2.2.3
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
